+++ example/lib/player/PlayingControls.dart
@@ -1,3 +1,4 @@
+import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
@@ -5,22 +6,71 @@ import '../asset_audio_player_icons.dart';
 
 class PlayingControls extends StatelessWidget {
   final bool isPlaying;
-  final bool isLooping;
+  final LoopMode loopMode;
   final bool isPlaylist;
   final Function() onPrevious;
   final Function() onPlay;
   final Function() onNext;
   final Function() toggleLoop;
+  final Function() onStop;
 
   PlayingControls({
     @required this.isPlaying,
     this.isPlaylist = false,
-    this.isLooping,
+    this.loopMode,
     this.toggleLoop,
     this.onPrevious,
     @required this.onPlay,
     this.onNext,
+    this.onStop,
   });
+
+  Widget _loopIcon(BuildContext context) {
+    final iconSize = 34.0;
+    if (loopMode == LoopMode.none) {
+      return Icon(
+        Icons.loop,
+        size: iconSize,
+        color: Colors.grey,
+      );
+    } else if (loopMode == LoopMode.playlist) {
+      return Icon(
+        Icons.loop,
+        size: iconSize,
+        color: Colors.black,
+      );
+    } else {
+      //single
+      return Stack(
+        alignment: Alignment.center,
+        children: [
+          Icon(
+            Icons.loop,
+            size: iconSize,
+            color: Colors.black,
+          ),
+          Center(
+            child: Text("1", style: TextStyle(fontSize: 12, fontWeight: FontWeight.bold),),
+          ),
+        ],
+      );
+    }
+    return NeumorphicRadio(
+      style: NeumorphicRadioStyle(
+        boxShape: NeumorphicBoxShape.circle(),
+      ),
+      padding: EdgeInsets.all(12),
+      value: LoopMode.playlist,
+      groupValue: this.loopMode,
+      child: Icon(
+        Icons.loop,
+        size: 18,
+      ),
+      onChanged: (newValue) {
+        toggleLoop();
+      },
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -28,20 +78,11 @@ class PlayingControls extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.center,
       mainAxisSize: MainAxisSize.max,
       children: [
-        NeumorphicRadio(
-          style: NeumorphicRadioStyle(
-            boxShape: NeumorphicBoxShape.circle(),
-          ),
-          padding: EdgeInsets.all(12),
-          value: true,
-          groupValue: this.isLooping,
-          child: Icon(
-            Icons.loop,
-            size: 18,
-          ),
-          onChanged: (newValue) {
+        GestureDetector(
+          onTap: () {
             toggleLoop();
           },
+          child: _loopIcon(context),
         ),
         SizedBox(
           width: 12,
@@ -64,9 +105,7 @@ class PlayingControls extends StatelessWidget {
           padding: EdgeInsets.all(24),
           onPressed: this.onPlay,
           child: Icon(
-            isPlaying
-                ? AssetAudioPlayerIcons.pause
-                : AssetAudioPlayerIcons.play,
+            isPlaying ? AssetAudioPlayerIcons.pause : AssetAudioPlayerIcons.play,
             size: 32,
           ),
         ),
@@ -83,7 +122,19 @@ class PlayingControls extends StatelessWidget {
         ),
         SizedBox(
           width: 45,
-        )
+        ),
+        if(onStop != null)
+          NeumorphicButton(
+            style: NeumorphicStyle(
+              boxShape: NeumorphicBoxShape.circle(),
+            ),
+            padding: EdgeInsets.all(16),
+            onPressed: this.onStop,
+            child: Icon(
+              AssetAudioPlayerIcons.stop,
+              size: 32,
+            ),
+          ),
       ],
     );
   }